+++ CHANGELOG.md
@@ -1,14 +1,28 @@
 ### [Report an Issue](https://github.com/atlassian/atlascode/issues)
 
+## What's new in 4.0.10
+
+### Features
+
+- Added development field support
+- Added Branch creation and Rovo Dev actions in Create Issue Page
+- Added filter by project support
+
+## What's new in 4.0.9
+
+### Features
+
+- Added filter by assignee support
+
 ## What's new in 4.0.8
 
 ### Features
 
 - Added activity history support
 
-### Bug Fixes 
+### Bug Fixes
 
-- Fixed issue with project loading in Jira Cloud Create Issue Screen 
+- Fixed issue with project loading in Jira Cloud Create Issue Screen
 
 ## What's new in 4.0.7